@@ -1,57 +1,133 @@
 apply plugin: 'com.android.application'
-apply plugin: 'android-apt'
+apply plugin: 'org.greenrobot.greendao'
+apply plugin: 'com.jakewharton.butterknife'
+apply plugin: 'com.hujiang.android-aspectjx'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.fb.sdclibrary"
         minSdkVersion 14
-        targetSdkVersion 25
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+
+        multiDexEnabled true
+
+        flavorDimensions "versionCode"
+
+        ndk {
+            //设置支持的SO库架构（开发者可以根据需要，选择一个或多个平台的so）, "armeabi-v7a", "arm64-v8a", "x86", "arm64-v8a", "x86_64"
+            abiFilters "armeabi"
+        }
     }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+
     buildTypes {
         release {
+            buildConfigField "String", "RECEIVER_EMAIL", "\"${RECEIVER_EMAIL}\""  //发送崩溃日志
+            buildConfigField "String", "SENDER_USER", "\"${SENDER_USER}\""
+            buildConfigField "String", "SENDER_PASSWORD", "\"${SENDER_PASSWORD}\""
+            buildConfigField "String", "SENDER_EMAIL", "\"${SENDER_EMAIL}\""
+            buildConfigField "String", "SMTP_HOST", "\"${SMTP_HOST}\""
+            buildConfigField "String", "SMTP_PORT", "\"${SMTP_PORT}\""
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+
+            applicationVariants.all { variant ->
+                variant.outputs.all { output ->
+                    def outputFile = output.outputFile
+                    def fileName
+                    if (outputFile != null && outputFile.name.endsWith('.apk')) {
+                        // 输出apk名称为fb_xfezt_1.0.0_2.apk
+                        if (variant.buildType.name == 'release') {
+                            fileName = "fb_family_well_${defaultConfig.versionName}_${defaultConfig.versionCode}_${variant.productFlavors[0].name}.apk"
+                        } else if (variant.buildType.name == 'debug') {
+                            fileName = "fb_family_well_debug.apk"
+                        } else {
+                            fileName = "fb_family_well_preview.apk"
+                        }
+                        outputFileName = fileName
+                    }
+                }
+            }
+        }
+        debug {
+            buildConfigField "String", "RECEIVER_EMAIL", "\"${RECEIVER_EMAIL}\""  //发送崩溃日志
+            buildConfigField "String", "SENDER_USER", "\"${SENDER_USER}\""
+            buildConfigField "String", "SENDER_PASSWORD", "\"${SENDER_PASSWORD}\""
+            buildConfigField "String", "SENDER_EMAIL", "\"${SENDER_EMAIL}\""
+            buildConfigField "String", "SMTP_HOST", "\"${SMTP_HOST}\""
+            buildConfigField "String", "SMTP_PORT", "\"${SMTP_PORT}\""
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+
+    greendao {
+        schemaVersion 1
+        daoPackage 'com.fb.sdclibrary.db.greendao'
+        targetGenDir 'src/main/java'
+    }
+
+    //移除lint检测的error
+    lintOptions {
+        abortOnError false
+        disable "ResourceType"
+    }
+    productFlavors {
+        xiaomi {}
+    }
+    productFlavors.all { flavor ->
+        flavor.manifestPlaceholders = [UMENG_CHANNEL_VALUE: name]
+    }
+    dexOptions {
+//        incremental true
+        javaMaxHeapSize "4g"
+    }
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    androidTestImplementation('androidx.test.espresso:espresso-core:3.1.0', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
+    testImplementation 'junit:junit:4.12'
 
-    compile 'com.android.support:appcompat-v7:25.3.1'
-    compile 'com.android.support:cardview-v7:25.3.1'
-    compile 'com.android.support:recyclerview-v7:25.3.1'
-    compile 'com.android.support.constraint:constraint-layout:1.0.2'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'androidx.multidex:multidex:2.0.1'
 
-    compile 'com.squareup.retrofit2:retrofit:2.1.0'
-    compile 'com.squareup.retrofit2:converter-gson:2.1.0'
-    compile 'com.squareup.retrofit2:adapter-rxjava:2.1.0'
-    compile 'io.reactivex:rxjava:1.2.3'
-    compile 'io.reactivex:rxandroid:1.2.1'
-    compile 'com.jakewharton:butterknife:8.8.1'
-    compile 'org.greenrobot:eventbus:3.1.1'
-    compile 'com.squareup.okhttp3:logging-interceptor:3.5.0'
-    compile 'org.greenrobot:greendao:3.1.0'
-    compile 'org.greenrobot:greendao-generator:3.1.0'
-    compile 'com.github.wenmingvs:LogReport:1.0.3'
-    compile 'com.github.bumptech.glide:glide:3.7.0'
-    compile 'com.github.razerdp:BasePopup:1.9.4'
-    compile 'cn.qqtheme.framework:WheelPicker:1.5.6'
-    compile 'com.skyfishjy.ripplebackground:library:1.0.1'
-    compile 'com.just.agentweb:agentweb:2.0.1'
-    compile 'com.github.PhilJay:MPAndroidChart:v3.0.1'
-    compile 'com.nineoldandroids:library:2.4.0'
-    compile 'com.github.GrenderG:Toasty:1.3.0'
-    testCompile 'junit:junit:4.12'
-    apt 'com.jakewharton:butterknife-compiler:8.8.1'
-    debugCompile 'com.squareup.leakcanary:leakcanary-android:1.3.1'
-    releaseCompile 'com.squareup.leakcanary:leakcanary-android-no-op:1.3.1'
+    implementation 'com.squareup.retrofit2:retrofit:2.3.0'
+    implementation 'com.squareup.retrofit2:converter-gson:2.3.0'
+    implementation 'com.squareup.retrofit2:adapter-rxjava:2.3.0'
+    implementation 'io.reactivex:rxjava:1.3.0'
+    implementation 'io.reactivex:rxandroid:1.2.1'
+    implementation 'org.greenrobot:eventbus:3.1.1'
+    implementation 'com.squareup.okhttp3:logging-interceptor:3.8.0'
+    implementation 'org.greenrobot:greendao:3.2.2'
+    implementation 'org.greenrobot:greendao-generator:3.2.2'
+    implementation 'com.github.wenmingvs:LogReport:1.0.3'
+    implementation 'com.github.bumptech.glide:glide:3.7.0'
+    implementation 'com.github.razerdp:BasePopup:1.9.4'
+    implementation 'cn.qqtheme.framework:WheelPicker:1.5.6'
+    implementation 'com.skyfishjy.ripplebackground:library:1.0.1'
+    implementation 'com.just.agentweb:agentweb:2.0.1'
+    implementation 'com.github.PhilJay:MPAndroidChart:v3.0.1'
+    implementation 'com.nineoldandroids:library:2.4.0'
+    implementation 'com.github.GrenderG:Toasty:1.3.0'
+    //implementation 'com.github.sendtion:LogReport:1.0.4'
+    implementation 'io.github.isliqian:NiceEmail:3.0.0'
+
+    implementation 'com.jakewharton:butterknife:10.1.0'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:10.1.0'
+    debugImplementation 'com.squareup.leakcanary:leakcanary-android:1.5'
+    releaseImplementation 'com.squareup.leakcanary:leakcanary-android-no-op:1.5.1'
 }
