@@ -1,8 +1,8 @@
 package com.fb.sdclibrary.activity;
 
 import android.content.pm.ActivityInfo;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AppCompatActivity;
+import androidx.core.content.ContextCompat;
+import androidx.appcompat.app.AppCompatActivity;
 import android.widget.Toast;
 
 import com.fb.sdclibrary.R;
