@@ -2,7 +2,7 @@ package com.fb.sdclibrary.service;
 
 import android.app.IntentService;
 import android.content.Intent;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.Log;
 
 import java.io.BufferedInputStream;
