@@ -1,7 +1,7 @@
 package com.fb.sdclibrary.utils;
 
 import android.app.Activity;
-import android.support.v4.content.ContextCompat;
+import androidx.core.content.ContextCompat;
 
 import com.fb.sdclibrary.R;
 
