@@ -6,8 +6,8 @@ import android.content.res.Resources;
 import android.graphics.Color;
 import android.os.Build;
 import android.os.Environment;
-import android.support.annotation.ColorInt;
-import android.support.v4.view.ViewCompat;
+import androidx.annotation.ColorInt;
+import androidx.core.view.ViewCompat;
 import android.util.AttributeSet;
 import android.util.DisplayMetrics;
 import android.view.Display;
